+++ cli/bin/index.ts
@@ -1,32 +1,14 @@
 #!/usr/bin/env node
-import fs from "fs";
-import { fileURLToPath } from "url";
 import { Command } from "commander";
-import { initConfig } from "./init/init.js";
-import { addComponent } from "./addComponent.js";
-import { addHook } from "./addHooks.js";
-import { updateInstalledComponents } from "./update.js";
-import { addUtil } from "./addUtils.js";
-import { addProvider } from "./addProvider.js";
-import { addLayout } from "./addLayout.js";
+import { add } from "../src/commands/add.js";
+import { initConfig } from "../src/commands/init.js";
+import { addHook } from "../src/commands/hook.js";
+import { addLayout } from "../src/commands/layout.js";
+import { addUtil } from "../src/commands/utils.js";
+import { addProvider } from "../src/commands/provider.js";
+import { updateInstalledComponents } from "../src/commands/update.js";
 
 const program = new Command();
-const __filename = fileURLToPath(import.meta.url);
-const CONFIG_FILE = "glare.json";
-
-export function getConfig() {
-  if (!fs.existsSync(CONFIG_FILE)) {
-    console.error('❌ glare.json not found. Run "npx torch-glare@latest init" first');
-    process.exit(1);
-  }
-
-  try {
-    return JSON.parse(fs.readFileSync(CONFIG_FILE, "utf8"));
-  } catch (error) {
-    console.error("❌ Error reading glare.json:", error.message);
-    process.exit(1);
-  }
-}
 
 program
   .name("torch-glare")
@@ -36,12 +18,12 @@ program
 program
   .command("init")
   .description("Initialize torch.json configuration file")
-  .action(() => initConfig(CONFIG_FILE));
+  .action(() => initConfig());
 
 program
   .command("add [component]")
   .description("Add a component interactively or install a specified one")
-  .action((component) => addComponent(component && `${component}.tsx`));
+  .action((component) => add(component && `${component}.tsx`));
 
 program
   .command("hook [hook]")
@@ -69,5 +51,3 @@ program
   .action(() => updateInstalledComponents());
 
 program.parse(process.argv);
-
-export { CONFIG_FILE, __filename };