+++ cli/src/commands/provider.ts
@@ -1,23 +1,29 @@
 import fs from "fs";
 import path from "path";
-import { getConfig } from "./cli.js";
+import { getConfig } from "../shared/getConfig.js";
+import { CONFIG_FILE } from "./init.js";
+import { Config } from "../types/main.js";
 import { fileURLToPath } from "url";
-import { ensureDirectoryExists, getComponentPaths, copyComponent } from "./addComponent.js";
+import { ensureDirectoryExists } from "../shared/ensureDirectoryExists.js";
+import { getInstallPaths } from "../shared/getInstallPaths.js";
+import { copyComponentsRecursively } from "../shared/copyComponentsRecursively.js";
 import inquirer from "inquirer";
+import { isFileExists } from "../shared/isFileExists.js";
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
 
 // Define the path to the provider templates directory
-const providerTemplatesDir = path.resolve(__dirname, "../../lib/providers");
+const providerTemplatesDir: string = path.resolve(__dirname, "../../../lib/providers");
 
 /**
  * Main function to add a provider and its dependencies.
  * @param {string} provider - The name of the provider to add.
+* @param {boolean} replace - Whether to replace the existing provider.
  */
-export async function addProvider(provider) {
-    const config = getConfig();
-    const availableProviders = getAvailableProviders(providerTemplatesDir);
+export async function addProvider(provider?: string, replace: boolean = false): Promise<void> {
+    const targetFile = getConfig(CONFIG_FILE) as Config;
+    const availableProviders: string[] = getAvailableProviders(providerTemplatesDir);
 
     // If no provider is provided, prompt the user to select one
     if (!provider) {
@@ -31,17 +37,20 @@ export async function addProvider(provider) {
     }
 
     // Get the path and create the target directory
-    const { source, targetDir } = getComponentPaths(provider, config, providerTemplatesDir, "providers");
-    const target = path.join(targetDir, provider);
-    fs.rmSync(target, { recursive: true, force: true });
-
+    const { source, targetDir } = getInstallPaths(provider, targetFile, providerTemplatesDir, "providers");
     // Ensure the target directory exists
     ensureDirectoryExists(targetDir);
 
-    // Copy the provider (file or directory) and install dependencies
-    copyComponent(source, target, addProvider);
+    // Check if provider already exists
+    if (isFileExists(targetDir) && !replace) {
+        console.log(`⚠️ Provider "${provider}" already exists.`);
+        return;
+    }
 
-    console.log(`✅ ${provider} has been added to ${config.path}!`);
+    // Copy the provider (file or directory) and install dependencies
+    copyComponentsRecursively(source, targetDir);
+
+    console.log(`✅ ${provider} has been added to ${targetFile.path}!`);
 }
 
 /**
@@ -49,7 +58,7 @@ export async function addProvider(provider) {
  * @param {string} providerTemplatesDir - Path to the provider templates directory.
  * @returns {string[]} - Array of provider names.
  */
-function getAvailableProviders(providerTemplatesDir) {
+function getAvailableProviders(providerTemplatesDir: string): string[] {
     return fs.readdirSync(providerTemplatesDir).map((file) => path.basename(file));
 }
 
@@ -58,7 +67,7 @@ function getAvailableProviders(providerTemplatesDir) {
  * @param {string[]} availableProviders - Array of available providers.
  * @returns {string} - The selected provider.
  */
-async function promptProviderSelection(availableProviders) {
+async function promptProviderSelection(availableProviders: string[]): Promise<string> {
     const { selectedProvider } = await inquirer.prompt([
         {
             type: "list",